+++ api/bootstrap/load_modules.R
@@ -143,6 +143,7 @@ bootstrap_load_modules <- function() {
     "functions/analysis-phenotype-mca-prep.R",
     "functions/analysis-phenotype-functions.R",
     "functions/analysis-null-models.R",
+    "functions/analysis-phenotype-missingness.R",
     "functions/analysis-cluster-validation.R",
     "functions/analysis-network-layout-functions.R",
     "functions/analysis-network-functions.R",